build: Add support for faking platform security patch level
SafetyNet will fail if the fake build fingerprints don't match the
platform security patch level, which is a problem for devices that are
no longer supported by the manufacturer.

To fix this issue, allow legacy products to define a fake security patch
level from their last official update and set the standard patch level
property to that. The real patch level is moved to another custom
property that is shown to the user, e.g. in Settings -> About.

Note that this requires the "core: Use real security patch level
property" commit in frameworks/base for users to see the correct patch
level in Settings.

Change-Id: I296185a54fe1f7da9faaa13eef35f3db4231daf0
Signed-off-by: Anushek Prasal <anushekprasal@gmail.com>
Signed-off-by: Aston-Martinn <advaithbhat9@gmail.com>